+++ src/nl/hannahsten/texifyidea/action/skim/ConfigureInverseSearchAction.kt
@@ -3,7 +3,7 @@ package nl.hannahsten.texifyidea.action.skim
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import nl.hannahsten.texifyidea.run.linuxpdfviewer.InternalPdfViewer
+import nl.hannahsten.texifyidea.run.pdfviewer.SkimViewer
 import nl.hannahsten.texifyidea.ui.SkimConfigureInverseSearchDialog
 
 /**
@@ -19,7 +19,7 @@ class ConfigureInverseSearchAction : AnAction() {
      * Hide this option when Skim is not available.
      */
     override fun update(e: AnActionEvent) {
-        e.presentation.isEnabledAndVisible = InternalPdfViewer.SKIM.isAvailable()
+        e.presentation.isEnabledAndVisible = SkimViewer.isAvailable()
     }
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT