+++ src/nl/hannahsten/texifyidea/run/pdfviewer/PdfViewer.kt
@@ -2,6 +2,8 @@ package nl.hannahsten.texifyidea.run.pdfviewer
 
 import com.intellij.openapi.extensions.ExtensionPointName
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.SystemInfo
+import nl.hannahsten.texifyidea.util.runCommand
 
 /**
  * Allow other plugins to define their own pdf viewers.
@@ -12,49 +14,103 @@ import com.intellij.openapi.project.Project
 private val EP_NAME = ExtensionPointName<ExternalPdfViewer>("nl.rubensten.texifyidea.pdfViewer")
 
 /**
- * Interface that defines a pdf viewer so we can use both [nl.hannahsten.texifyidea.run.linuxpdfviewer.InternalPdfViewer]s and [ExternalPdfViewer]s interchangeably.
+ * Interface that defines the general behavior of a PDF viewer.
  */
 interface PdfViewer {
 
+    /**
+     * The identifier of the viewer.
+     */
     val name: String?
     val displayName: String?
 
+    /**
+     * Tells whether the PDF viewer is available.
+     */
     fun isAvailable(): Boolean
 
     /**
-     * Performs a forward search to navigate from the source file to the corresponding location in the PDF document.
-     *
-     * @param pdfPath The file path to the PDF document. If null, the viewer may attempt to determine the path automatically.
-     * @param sourceFilePath The file path to the source file (e.g., a LaTeX file) that corresponds to the PDF document.
-     * @param line The line number in the source file to which the PDF viewer should navigate.
-     * @param project The project context in which the forward search is performed.
-     * @param focusAllowed Indicates whether the PDF viewer is allowed to take focus when performing the forward search. If false, the viewer may remain in the background.
+     * Whether the PDF viewer supports forward search.
+     * Default is true for backwards compatibility with the PDF Viewer plugin
      */
-    fun forwardSearch(pdfPath: String?, sourceFilePath: String, line: Int, project: Project, focusAllowed: Boolean)
+    val isForwardSearchSupported: Boolean
+        get() = true
+
+    /**
+     * Whether the PDF viewer supports customized focus change during forward search, namely, it can determine to take focus or not when performing forward search.
+     */
+    val isFocusSupported: Boolean
+        get() = false
+
+    /**
+     * Performs a forward search to navigate from the source file to the corresponding location in the document.
+     *
+     * @param outputPath The file path to the output document. If null, the viewer may attempt to determine the path automatically.
+     * @param sourceFilePath The file path to the source file (e.g., a LaTeX file) that corresponds to the document.
+     * @param line The line number in the source file to which the viewer should navigate.
+     * @param project The project context in which the forward search is performed.
+     * @param focusAllowed Indicates whether the viewer is allowed to take focus when performing the forward search.
+     * **If false, the viewer should not take focus anyway, even if it has to abort the operation.** (In order to not disturb the user.)
+     */
+    fun forwardSearch(outputPath: String?, sourceFilePath: String, line: Int, project: Project, focusAllowed: Boolean) {
+        // Default implementation does nothing.
+    }
 
     /**
      * Opens a PDF file in the viewer.
      *
      * @param pdfPath The file path to the PDF document that should be opened.
      * @param project The project context associated with the PDF file.
-     * @param newWindow Indicates whether the PDF should be opened in a new window. Defaults to false.
-     * @param focus Indicates whether the viewer should take focus when opening the file. Defaults to false.
-     * @param forceRefresh Indicates whether the viewer should force a refresh of the file content. Defaults to false.
+     * @param newWindow Indicates whether the PDF should be opened in a new window.
+     * @param focusAllowed Indicates whether the viewer should take focus when opening the file.
+     * **If false, the viewer should not take focus anyway, even if it has to abort the operation.** (In order to not disturb the user.)
+     * @param forceRefresh Indicates whether the viewer should force a refresh of the file content.
      */
-    fun openFile(pdfPath: String, project: Project, newWindow: Boolean = false, focus: Boolean = false, forceRefresh: Boolean = false) {
+    fun openFile(pdfPath: String, project: Project, newWindow: Boolean = false, focusAllowed: Boolean = true, forceRefresh: Boolean = false) {
         // Default implementation does nothing.
+    }
+
+    companion object {
+
+        val internalViewers
+            get() = listOf(SumatraViewer, EvinceViewer, OkularViewer, ZathuraViewer, SkimViewer, SystemDefaultViewer, NoViewer)
+
+        val externalViewers
+            get() = EP_NAME.extensionList
+
+        /**
+         * Gets the list of all PDF viewers, both internal and external.
+         */
+        val allViewers: List<PdfViewer>
+            get() = externalViewers + internalViewers
+
+        /**
+         * Gets the list of all available PDF viewers, both internal and external.
+         */
+        val availableViewers: List<PdfViewer>
+            get() = allViewers.filter { it.isAvailable() }
+
+        val firstAvailableViewer: PdfViewer
+            get() {
+                if (SystemInfo.isLinux) {
+                    // Use system default if possible
+                    return runCommand("xdg-mime", "query", "default", "application/pdf", timeout = 1)?.let { mime ->
+                        availableViewers.firstOrNull { viewer ->
+                            val name = viewer.name ?: return@firstOrNull false
+                            name.lowercase() in mime.lowercase()
+                        }
+                    } ?: availableViewers.first()
+                }
+                else {
+                    return allViewers.first { it.isAvailable() }
+                }
+            }
     }
 }
 
 /**
  * Define behaviour that external pdf viewers should inherit.
+ *
+ * This interface is kept for backwards compatibility.
  */
 interface ExternalPdfViewer : PdfViewer
-
-/**
- * Define functions that handle all external pdf viewers one by one.
- */
-object ExternalPdfViewers {
-
-    fun getExternalPdfViewers(): List<ExternalPdfViewer> = EP_NAME.extensionList
-}