+++ src/nl/hannahsten/texifyidea/util/CommandRunner.kt
@@ -5,8 +5,8 @@ import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.util.io.awaitExit
 import kotlinx.coroutines.*
 import java.io.BufferedReader
-import java.io.File
 import java.io.IOException
+import java.nio.file.Path
 import javax.swing.SwingUtilities
 
 data class CommandResult(
@@ -30,7 +30,7 @@ data class CommandResult(
  */
 suspend fun runCommandNonBlocking(
     vararg commands: String,
-    workingDirectory: File? = null,
+    workingDirectory: Path? = null,
     input: String? = null,
     discardOutput: Boolean = false,
     returnExceptionMessageAsErrorOutput: Boolean = false,
@@ -41,15 +41,15 @@ suspend fun runCommandNonBlocking(
 
         // where/which commands occur often but do not change since the output depends on PATH, so can be cached
         val isExecutableLocationCommand = commands.size == 2 && listOf("where", "which").contains(commands[0])
-        if (isExecutableLocationCommand && SystemEnvironment.executableLocationCache[commands[1]] != null) {
-            val standardOutput = SystemEnvironment.executableLocationCache[commands[1]]
-            Log.debug("Retrieved output of $commands from cache: $standardOutput")
-            return@withContext CommandResult(0, standardOutput, null)
+        val cachedExecutableLocation = SystemEnvironment.executableLocationCache[commands[1]]
+        if (isExecutableLocationCommand && cachedExecutableLocation != null) {
+            Log.debug("Retrieved output of $commands from cache: $cachedExecutableLocation")
+            return@withContext CommandResult(0, cachedExecutableLocation, null)
         }
 
         val processBuilder = GeneralCommandLine(*commands)
             .withParentEnvironmentType(GeneralCommandLine.ParentEnvironmentType.CONSOLE)
-            .withWorkDirectory(workingDirectory)
+            .apply { if (workingDirectory != null) withWorkingDirectory(workingDirectory) }
             .toProcessBuilder()
 
         if (discardOutput) {
@@ -130,7 +130,7 @@ catch (e: IOException) {
  *
  * @return The output of the command or null if an exception was thrown.
  */
-fun runCommand(vararg commands: String, workingDirectory: File? = null, timeout: Long = 3): String? =
+fun runCommand(vararg commands: String, workingDirectory: Path? = null, timeout: Long = 3): String? =
     runBlocking {
         runCommandNonBlocking(*commands, workingDirectory = workingDirectory, timeout = timeout).output
     }
@@ -144,7 +144,7 @@ fun runCommand(vararg commands: String, workingDirectory: File? = null, timeout:
  */
 fun runCommandWithExitCode(
     vararg commands: String,
-    workingDirectory: File? = null,
+    workingDirectory: Path? = null,
     timeout: Long = 3,
     returnExceptionMessage: Boolean = false,
     discardOutput: Boolean = false,