+++ src/nl/hannahsten/texifyidea/util/files/ReferencedFileSetCache.kt
@@ -3,14 +3,13 @@ package nl.hannahsten.texifyidea.util.files
 import arrow.atomic.AtomicBoolean
 import com.intellij.openapi.application.smartReadAction
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.platform.util.progress.ProgressReporter
 import com.intellij.psi.PsiFile
 import com.intellij.psi.SmartPsiElementPointer
 import com.intellij.psi.createSmartPointer
-import kotlinx.coroutines.sync.Mutex
 import nl.hannahsten.texifyidea.file.listeners.VfsChangeListener
 import nl.hannahsten.texifyidea.index.LatexIncludesIndex
+import nl.hannahsten.texifyidea.util.files.ReferencedFileSetCache.Cache.fileSetCache
 import nl.hannahsten.texifyidea.util.runInBackgroundNonBlocking
 import java.util.concurrent.ConcurrentHashMap
 
@@ -20,11 +19,6 @@ import java.util.concurrent.ConcurrentHashMap
  * @author Hannah Schellekens
  */
 class ReferencedFileSetCache {
-
-    companion object {
-
-        private val mutex = Mutex()
-    }
 
     object Cache {
 
@@ -40,12 +34,17 @@ class ReferencedFileSetCache {
          * For the same reason we do not use a CachedValue, because the CachedValuesManager is project-specific.
          *
          * We use SmartPsiElementPointer to avoid storing files which have become invalid, e.g. after installing a plugin which doesn't require a restart.
+         *
+         * We need to keep track per project, because a cache fill can only be done for a given project, so we want to avoid dropping caches for other projects.
          */
-        internal var fileSetCache = ConcurrentHashMap<String, Set<SmartPsiElementPointer<PsiFile>>>()
+        internal var fileSetCache = mutableMapOf<Project, MutableMap<String, Set<SmartPsiElementPointer<PsiFile>>>>()
 
-        internal var rootFilesCache = ConcurrentHashMap<String, Set<SmartPsiElementPointer<PsiFile>>>()
+        /**
+         * Similar to [fileSetCache], maps file paths to root files of the file set.
+         */
+        internal var rootFilesCache = ConcurrentHashMap<Project, MutableMap<String, Set<SmartPsiElementPointer<PsiFile>>>>()
 
-        internal var isCacheFillInProgress = AtomicBoolean(false)
+        internal var isCacheFillInProgress = mutableMapOf<Project, AtomicBoolean>()
 
         /**
          * The number of includes in the include index at the time the cache was last filled.
@@ -63,29 +62,20 @@ class ReferencedFileSetCache {
      */
     @Synchronized
     fun fileSetFor(file: PsiFile, useIndexCache: Boolean = true): Set<PsiFile> {
-        return getSetFromCache(file, Cache.fileSetCache, useIndexCache)
+        return getSetFromCache(file, fileSetCache.getOrPut(file.project) { mutableMapOf() }, useIndexCache)
     }
 
     @Synchronized
     fun rootFilesFor(file: PsiFile, useIndexCache: Boolean = true): Set<PsiFile> {
-        return getSetFromCache(file, Cache.rootFilesCache, useIndexCache)
-    }
-
-    /**
-     * Clears the cache for base file `file`.
-     * Note: this will not trigger a cache refill
-     */
-    fun dropCaches(file: VirtualFile) {
-        Cache.fileSetCache.remove(file.path)
-        Cache.rootFilesCache.remove(file.path)
+        return getSetFromCache(file, Cache.rootFilesCache.getOrPut(file.project) { mutableMapOf() }, useIndexCache)
     }
 
     /**
      * Note: this will not trigger a cache refill
      */
-    fun dropAllCaches() {
-        Cache.fileSetCache.clear()
-        Cache.rootFilesCache.clear()
+    fun dropAllCaches(project: Project) {
+        fileSetCache[project]?.clear()
+        Cache.rootFilesCache[project]?.clear()
     }
 
     /**
@@ -99,12 +89,12 @@ class ReferencedFileSetCache {
      * Since we have to calculate the fileset to fill the root file or fileset cache, we make sure to only do that
      * once and then fill both caches with all the information we have.
      */
-    private suspend fun getNewCachesFor(requestedFile: PsiFile, reporter: ProgressReporter?): Pair<Map<String, Set<SmartPsiElementPointer<PsiFile>>>, Map<String, Set<SmartPsiElementPointer<PsiFile>>>> {
+    private suspend fun getNewCachesFor(project: Project, reporter: ProgressReporter?): Pair<Map<String, Set<SmartPsiElementPointer<PsiFile>>>, Map<String, Set<SmartPsiElementPointer<PsiFile>>>> {
         val newFileSetCache = mutableMapOf<String, Set<SmartPsiElementPointer<PsiFile>>>()
         val newRootFilesCache = mutableMapOf<String, Set<SmartPsiElementPointer<PsiFile>>>()
 
-        val filesets = requestedFile.project.findReferencedFileSetWithoutCache(reporter).toMutableMap()
-        val tectonicInclusions = findTectonicTomlInclusions(requestedFile.project)
+        val filesets = project.findReferencedFileSetWithoutCache(reporter).toMutableMap()
+        val tectonicInclusions = findTectonicTomlInclusions(project)
 
         // Now we join all the file sets that are in the same file set according to the Tectonic.toml file
         for (inclusionsSet in tectonicInclusions) {
@@ -117,7 +107,7 @@ class ReferencedFileSetCache {
 
         for (fileset in filesets.values) {
             val validFiles = fileset.mapNotNull {
-                smartReadAction(requestedFile.project) {
+                smartReadAction(project) {
                     if (it.isValid) it.createSmartPointer() else null
                 }
             }.toSet()
@@ -125,9 +115,17 @@ class ReferencedFileSetCache {
                 newFileSetCache[file.virtualFile.path] = validFiles
             }
 
-            val rootFiles = requestedFile.findRootFilesWithoutCache(fileset)
+            // Find root files, either this or other files
+            val rootFiles = mutableSetOf<PsiFile>()
+            for (file in fileset) {
+                rootFiles.addAll(file.findRootFilesWithoutCache())
+                if (smartReadAction(project) { file.isRoot() }) {
+                    rootFiles.add(file)
+                }
+            }
+
             val validRootFiles = rootFiles.mapNotNull {
-                smartReadAction(requestedFile.project) {
+                smartReadAction(project) {
                     if (it.isValid) it.createSmartPointer() else null
                 }
             }.toSet()
@@ -142,37 +140,39 @@ class ReferencedFileSetCache {
     /**
      * In a thread-safe way, get the value from the cache and if needed refresh the cache first.
      */
-    private fun getSetFromCache(file: PsiFile, cache: ConcurrentHashMap<String, Set<SmartPsiElementPointer<PsiFile>>>, useIndexCache: Boolean = true): Set<PsiFile> {
+    private fun getSetFromCache(file: PsiFile, cache: MutableMap<String, Set<SmartPsiElementPointer<PsiFile>>>, useIndexCache: Boolean = true): Set<PsiFile> {
         if (file.virtualFile == null) {
             return setOf(file)
         }
+        val project = file.project
 
         // Use the keys of the whole project, because suppose a new include includes the current file, it could be anywhere in the project
         // Note that LatexIncludesIndex.Util.getItems(file.project) may be a slow operation and should not be run on EDT
-        val includes = LatexIncludesIndex.Util.getItems(file.project, useCache = useIndexCache)
+        // Don't use cache here, otherwise we would just be comparing cache with cache
+        val numberOfIncludes = LatexIncludesIndex.Util.getNumberOfIndexedItems(project)
 
         // The cache should be complete once filled, any files not in there are assumed to not be part of a file set that has a valid root file
-        if (includes.size != Cache.numberOfIncludes[file.project] && !Cache.isCacheFillInProgress.getAndSet(true)) {
-            Cache.numberOfIncludes[file.project] = includes.size
+        if (numberOfIncludes != Cache.numberOfIncludes[project] && !Cache.isCacheFillInProgress.getOrPut(project) { AtomicBoolean(false) }.getAndSet(true)) {
+            Cache.numberOfIncludes[project] = numberOfIncludes
 
-            runInBackgroundNonBlocking(file.project, "Updating file set cache...") { reporter ->
+            runInBackgroundNonBlocking(project, "Updating file set cache...") { reporter ->
                 try {
                     // Only drop caches after we have new data (since that task may be cancelled)
-                    val (newFileSetCache, newRootFilesCache) = getNewCachesFor(file, reporter)
-                    dropAllCaches()
-                    Cache.fileSetCache.putAll(ConcurrentHashMap(newFileSetCache.toMutableMap()))
-                    Cache.rootFilesCache.putAll(ConcurrentHashMap(newRootFilesCache.toMutableMap()))
+                    val (newFileSetCache, newRootFilesCache) = getNewCachesFor(project, reporter)
+                    dropAllCaches(project)
+                    fileSetCache.getOrPut(project) { mutableMapOf() }.putAll(newFileSetCache.toMutableMap())
+                    Cache.rootFilesCache.getOrPut(project) { mutableMapOf() }.putAll(newRootFilesCache.toMutableMap())
                     // Many inspections use the file set, so a rerun could give different results
-                    smartReadAction(file.project) { file.rerunInspections() }
+                    smartReadAction(project) { file.rerunInspections() }
                 }
                 finally {
-                    Cache.isCacheFillInProgress.set(false)
+                    Cache.isCacheFillInProgress[project]?.set(false)
                 }
             }
         }
 
         // Make sure to check if file is still valid after retrieving from cache (it may have been deleted)
         val fileset = cache[file.virtualFile.path]?.mapNotNull { it.element }?.filter { it.isValid }?.toSet()
-        return if (fileset.isNullOrEmpty())setOf(file) else fileset
+        return if (fileset.isNullOrEmpty()) setOf(file) else fileset
     }
 }