+++ src/nl/hannahsten/texifyidea/util/files/impl/ReferencedFileSetServiceImpl.kt
@@ -1,6 +1,6 @@
 package nl.hannahsten.texifyidea.util.files.impl
 
-import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 import nl.hannahsten.texifyidea.util.files.ReferencedFileSetCache
 import nl.hannahsten.texifyidea.util.files.ReferencedFileSetService
@@ -16,9 +16,7 @@ class ReferencedFileSetServiceImpl : ReferencedFileSetService {
 
     override fun rootFilesOf(psiFile: PsiFile, useIndexCache: Boolean): Set<PsiFile> = cache.rootFilesFor(psiFile, useIndexCache)
 
-    override fun dropCaches(file: VirtualFile) = cache.dropCaches(file)
-
-    override fun dropAllCaches() = cache.dropAllCaches()
+    override fun dropAllCaches(project: Project) = cache.dropAllCaches(project)
 
     override fun markCacheOutOfDate() = cache.markCacheOutOfDate()
 }